Nucleo improvements, add Resource concept

+++ src/Domains/Nucleo/Observer.php
@@ -3,6 +3,7 @@
 namespace SuperV\Platform\Domains\Nucleo;
 
 use Illuminate\Database\Eloquent\Model;
+use SuperV\Platform\Domains\Nucleo\Contracts\Structable;
 
 class Observer
 {
@@ -33,13 +34,14 @@ class Observer
             }
 
             if ($field->hasRules()) {
-                $rules[$field->slug] = $field->rules;
+                $rules[$field->slug] = $field->getRules();
                 $attributes[$field->slug] = sprintf('%s.%s', $model->getTable(), $field->slug);
                 $data[$field->slug] = $model->getAttribute($field->slug);
             }
 
             if ($field->scatter) {
-                $model->__cache[$field->slug] = $model->offsetGet($field->slug);  /** hate */
+                $model->__cache[$field->slug] = $model->offsetGet($field->slug);
+                /** hate */
                 $model->offsetUnset($field->slug);
             }
         }
@@ -50,26 +52,30 @@ class Observer
 
     public function created(Model $model)
     {
-        $struct = Struct::create(
-            [
-                'related_id'   => $model->id,
-                'prototype_id' => $model->prototype()->id,
-            ]
-        );
+        if ($model instanceof Structable) {
+            $struct = Struct::create(
+                [
+                    'related_id'   => $model->id,
+                    'prototype_id' => $model->prototype()->id,
+                ]
+            );
 
-        $model->fields()->map(function (Field $field) use ($struct, $model) {
-            $struct->members()->create(['field_id' => $field->id]);
-        });
+            $model->fields()->map(function (Field $field) use ($struct, $model) {
+                $struct->members()->create(['field_id' => $field->id]);
+            });
+        }
     }
 
     public function deleted(Model $model)
     {
-        $prototype = Prototype::where('slug', $model->getTable())->first();
+        if ($model instanceof Structable) {
+            $prototype = Prototype::where('slug', $model->getTable())->first();
 
-        $struct = $prototype->structs()->where('related_id', $model->id)->first();
+            $struct = $prototype->structs()->where('related_id', $model->id)->first();
 
-        if ($struct) {
-            $struct->delete();
+            if ($struct) {
+                $struct->delete();
+            }
         }
     }
 
@@ -84,10 +90,12 @@ class Observer
                 continue;
             }
 
-            $member = $model->struct()->member($field->slug);
+            if ($model instanceof Structable) {
+                $member = $model->struct()->member($field->slug);
 
-            $value = $field->scatter ? array_get($model->__cache  /** hate */, $field->slug) : $model->getAttribute($field->slug);
-            $member->setValue($value);
+                $value = $field->scatter ? array_get($model->__cache/** hate */, $field->slug) : $model->getAttribute($field->slug);
+                $member->setValue($value);
+            }
         }
     }
 }
